+++ src/app/(protected)/crisis-resources/index.tsx
@@ -1,0 +1,5 @@
+import { CrisisResourcesScreen } from '@/src/features/crisis-resources/crisis-resources-screen';
+
+export default function CrisisResourcesRoute() {
+  return <CrisisResourcesScreen />;
+}